refactor(infrastructure): implement proper dependency injection into all clients and managers, re-organize project into services, core, tools
