+++ supabase_mcp/services/api/specs/api_spec.json
