+++ supabase_mcp/services/database/sql/queries/get_schemas.sql
