+++ supabase_mcp/services/database/sql/queries/get_table_schema.sql
