+++ supabase_mcp/services/database/sql/queries/get_tables.sql
