+++ supabase_mcp/tools/__init__.py
@@ -3,6 +3,6 @@
 This module provides a centralized way to manage tool descriptions and registration.
 """
 
-from .tool_manager import ToolManager, ToolName
+from .manager import ToolManager, ToolName
 
 __all__ = ["ToolManager", "ToolName"]
