+++ tests/test_main.py
@@ -5,7 +5,7 @@ from unittest.mock import ANY, patch
 import pytest
 
 from supabase_mcp.logger import logger
-from supabase_mcp.main import inspector, mcp, run
+from supabase_mcp.main import mcp, run_inspector, run_server
 
 # === UNIT TESTS ===
 
@@ -55,14 +55,14 @@ class TestMain:
     def test_run_server_starts(self):
         """Test that server run function executes without errors"""
         with patch("supabase_mcp.main.mcp.run") as mock_run:
-            run()
+            run_server()
             mock_run.assert_called_once()
 
     @pytest.mark.unit
     def test_inspector_mode(self):
         """Test that inspector mode initializes correctly"""
         with patch("mcp.cli.cli.dev") as mock_dev:
-            inspector()
+            run_inspector()
             mock_dev.assert_called_once_with(file_spec=ANY)
 
     @pytest.mark.unit
